requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.7
pycodestyle==2.11.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
